+++ code/src/cljs/sixsq/nuvla/ui/apps_application/subs.cljs
@@ -27,7 +27,7 @@
   ::license-error?
   (fn [db]
     #_:clj-kondo/ignore
-    (not (empty? (::spec/license-validation-errors db)))))
+    (seq (::spec/license-validation-errors db))))
 
 
 (reg-sub